# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger API for Golang GoTest.",
        "title": "GoTest Swagger API",
        "contact": {
            "name": "Rovshan Gasanov",
            "email": "rovshan27121@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8181",
    "basePath": "/api/",
    "paths": {
        "/baskets/product": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Baskets"
                ],
                "summary": "Add product to basket",
                "parameters": [
                    {
                        "description": "new product added",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.BasketProduct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.BasketProduct"
                        }
                    }
                }
            }
        },
        "/baskets/product/{basket_id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Baskets"
                ],
                "summary": "Delete product to basket",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "BasketID",
                        "name": "basket_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/baskets/product/{product_id}/{basket_id}": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Baskets"
                ],
                "summary": "Decrease quantity product to basket",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ProductID",
                        "name": "product_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "BasketID",
                        "name": "basket_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "new product added",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.BasketProduct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.BasketProduct"
                        }
                    }
                }
            }
        },
        "/baskets/{basket_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Baskets"
                ],
                "summary": "Get basket",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "BasketID",
                        "name": "basket_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Basket"
                        }
                    }
                }
            }
        },
...
        "/orders/{order_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Get order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "OrderID",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Order"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Delete order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "OrderID",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/orders/{user_id}": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Add order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "UserID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Order"
                        }
                    }
                }
            }
        },
        "/products": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Product"
                            }
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Add product",
                "parameters": [
                    {
                        "description": "new product",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Product"
                        }
                    }
                }
            }
        },
        "/products/{product_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ProductID",
                        "name": "product_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Product"
                        }
                    }
                }
            },
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Edit product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ProductID",
                        "name": "product_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "edit product",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Product"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/jsondelete"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Delete product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ProductID",
                        "name": "product_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.Basket": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 0
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.BasketProduct"
                    }
                },
                "total_price": {
                    "type": "integer",
                    "example": 400
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "domain.BasketProduct": {
            "type": "object",
            "properties": {
                "basket_id": {
                    "type": "integer",
                    "example": 1
                },
                "count": {
                    "type": "integer",
                    "example": 2
                },
                "id": {
                    "type": "integer",
                    "example": 0
                },
                "product_id": {
                    "type": "integer",
                    "example": 1
                },
                "total_price": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
        "domain.Order": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "status_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "domain.Product": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "test description"
                },
                "id": {
                    "type": "integer",
                    "example": 0
                },
                "name": {
                    "type": "string",
                    "example": "testProduct"
                },
                "price": {
                    "type": "integer",
                    "example": 300
                }
            }
        },
...
    }
}